acess restriction added

+++ src/main/java/com/example/Eshop/repos/ProductRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface ProductRepo extends CrudRepository<Product, Long> {
+public interface ProductRepo extends CrudRepository<Product, Integer> {
     List<Product> findByTag(String tag);
 }
